# Use smaller fixture objects (fewer children) to speed up tests