--- a/spec/services/metadata_cloud_service_spec.rb
+++ b/spec/services/metadata_cloud_service_spec.rb
@@ -6,21 +6,26 @@ WebMock.allow_net_connect!
 
 RSpec.describe MetadataCloudService, prep_metadata_sources: true do
   let(:mcs) { described_class.new }
-  let(:oid) { "16371272" }
+  let(:oid) { "16371253" }
   let(:oid_url) { "https://#{described_class.metadata_cloud_host}/metadatacloud/api/ladybird/oid/#{oid}?mediaType=json" }
   let(:short_oid_path) { Rails.root.join("spec", "fixtures", "short_fixture_ids.csv") }
-
+  before do
+    stub_metadata_cloud("2034600")
+    stub_metadata_cloud("2005512")
+    stub_metadata_cloud("16414889")
+    stub_metadata_cloud("14716192")
+    stub_metadata_cloud("16854285")
+  end
   context "creating a ParentObject from an import" do
     before do
-      stub_request(:get, "https://metadata-api-test.library.yale.edu/metadatacloud/api/ladybird/oid/16371272")
-        .to_return(status: 200, body: File.open(File.join(fixture_path, "ladybird", "16371272.json")).read)
+      stub_metadata_cloud(16_371_253)
     end
     it "can create a parent_object from an array of oids" do
       expect(ParentObject.count).to eq 0
-      described_class.create_parent_objects_from_oids(["16371272"], "ladybird")
+      described_class.create_parent_objects_from_oids(["16371253"], "ladybird")
       expect(ParentObject.count).to eq 1
-      expect(ParentObject.where(oid: "16371272").first.ladybird_json).not_to be nil
-      expect(ParentObject.where(oid: "16371272").first.ladybird_json).not_to be_empty
+      expect(ParentObject.where(oid: "16371253").first.ladybird_json).not_to be nil
+      expect(ParentObject.where(oid: "16371253").first.ladybird_json).not_to be_empty
     end
   end
 
@@ -39,7 +44,13 @@ RSpec.describe MetadataCloudService, prep_metadata_sources: true do
   context "saving a Voyager record" do
     let(:path_to_example_file) { Rails.root.join("spec", "fixtures", "ils", "V-2034600.json") }
     let(:metadata_source) { "ils" }
-
+    before do
+      stub_metadata_cloud("V-2034600", "ils")
+      stub_metadata_cloud("V-2005512", "ils")
+      stub_metadata_cloud("V-16414889", "ils")
+      stub_metadata_cloud("V-14716192", "ils")
+      stub_metadata_cloud("V-16854285", "ils")
+    end
     it "can pull voyager records" do
       time_stamp_before = File.mtime(path_to_example_file.to_s)
       described_class.refresh_fixture_data(short_oid_path, metadata_source)
@@ -55,7 +66,10 @@ RSpec.describe MetadataCloudService, prep_metadata_sources: true do
 
     let(:path_to_example_file) { Rails.root.join("spec", "fixtures", "aspace", "AS-16854285.json") }
     let(:metadata_source) { "aspace" }
-
+    before do
+      stub_metadata_cloud("AS-16854285", "aspace")
+      # stub_metadata_cloud("AS-2034600", "aspace")
+    end
     it "can pull ArchiveSpace records" do
       time_stamp_before = File.mtime(path_to_example_file.to_s)
       described_class.refresh_fixture_data(short_oid_path, metadata_source)
@@ -102,6 +116,12 @@ RSpec.describe MetadataCloudService, prep_metadata_sources: true do
   context "if the MetadataCloud cannot find an object" do
     let(:unfindable_oid_array) { ["17063396", "17029210"] }
     let(:path_to_example_file) { Rails.root.join("spec", "fixtures", "ladybird", "17063396.json") }
+    around do |example|
+      original_vpn = ENV['VPN']
+      ENV['VPN'] = 'true'
+      example.run
+      ENV['VPN'] = original_vpn
+    end
     before do
       stub_request(:get, "https://#{described_class.metadata_cloud_host}/metadatacloud/api/ladybird/oid/17063396?include-children=1")
         .to_return(status: 400, body: "ex: can't connect to ladybird")