--- a/spec/spec_helper.rb
+++ b/spec/spec_helper.rb
@@ -34,7 +34,8 @@ require 'webmock/rspec'
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
   # Allow connections to local services / localhost
-  WebMock.disable_net_connect!(allow: "solr")
+  allowed_sites = ["solr", "localhost"]
+  WebMock.disable_net_connect!(allow: allowed_sites)
 
   # rspec-expectations config goes here. You can use an alternate
   # assertion/expectation library such as wrong or the stdlib/minitest