… spec/fixtures/short_fixture_ids.csv
… spec/fixtures/short_fixture_ids.pdf