# Before
```
Top 10 slowest examples (119.55 seconds, 62.3% of total time):
  Oid CSV Import when uploading a csv uploads and increases csv count
    14.24 seconds ./spec/system/csv_import_spec.rb:15
  MetadataCloudService saving a Voyager record can pull voyager records
    12.67 seconds ./spec/services/metadata_cloud_service_spec.rb:43
  MetadataCloudService saving an ArchiveSpace record can pull ArchiveSpace records
    12.05 seconds ./spec/services/metadata_cloud_service_spec.rb:59
  ActivityStreamReader daily automated updates marks objects as updated in the database
    11.62 seconds ./spec/lib/activity_stream_reader_spec.rb:247
  ActivityStreamReader daily automated updates processes the entire activity stream if it has never been run before
    11.58 seconds ./spec/lib/activity_stream_reader_spec.rb:278
  MetadataCloudService it gets called from a rake task is easy to invoke
    11.55 seconds ./spec/services/metadata_cloud_service_spec.rb:31
  Oid CSV Import when uploading a csv Gives a success message
    11.53 seconds ./spec/system/csv_import_spec.rb:22
  ActivityStreamReader daily automated updates can process the partial activity stream if there is a previous successful run
    11.52 seconds ./spec/lib/activity_stream_reader_spec.rb:266
  JsonFile can save a ParentObject to json
    11.43 seconds ./spec/models/concerns/json_file_spec.rb:11
  ActivityStreamReader with multiple oids for a single dependent uri creating a set of dependent_uris from database creates a set of local dependent uris
    11.36 seconds ./spec/lib/activity_stream_reader_spec.rb:393

Top 10 slowest example groups:
  Oid CSV Import
    12.89 seconds average (25.77 seconds / 2 examples) ./spec/system/csv_import_spec.rb:4
  JsonFile
    11.43 seconds average (11.43 seconds / 1 example) ./spec/models/concerns/json_file_spec.rb:4
  ActivityStreamReader
    5.34 seconds average (80.08 seconds / 15 examples) ./spec/lib/activity_stream_reader_spec.rb:5
  MetadataCloudService
    3.54 seconds average (38.9 seconds / 11 examples) ./spec/services/metadata_cloud_service_spec.rb:7
  OidImport
    2.54 seconds average (10.18 seconds / 4 examples) ./spec/models/oid_import_spec.rb:5
  ParentObject
    2.19 seconds average (10.94 seconds / 5 examples) ./spec/models/parent_object_solr_spec.rb:5
  ParentObjects
    0.34889 seconds average (4.54 seconds / 13 examples) ./spec/system/parent_object_spec.rb:4
  METS Xml Imports
    0.1916 seconds average (0.5748 seconds / 3 examples) ./spec/system/mets_xml_imports_spec.rb:4
  IiifManifestFactory
    0.16248 seconds average (3.09 seconds / 19 examples) ./spec/lib/iiif_manifest_factory_spec.rb:5
  /child_objects
    0.13498 seconds average (1.75 seconds / 13 examples) ./spec/requests/child_objects_spec.rb:16

Finished in 3 minutes 11.9 seconds (files took 2.94 seconds to load)
```

# After
```Top 10 slowest examples (11.27 seconds, 41.1% of total time):
  MetadataCloudService saving an ArchiveSpace record can pull ArchiveSpace records
    1.65 seconds ./spec/services/metadata_cloud_service_spec.rb:73
  METS Xml Imports when uploading a METS xml uploads and increases MetsImport and ParentObject count
    1.54 seconds ./spec/system/mets_xml_imports_spec.rb:10
  ParentObject indexing to Solr from the database with Ladybird ParentObjects can index the 5 parent objects in the database to Solr
    1.32 seconds ./spec/models/parent_object_solr_spec.rb:7
  Oid CSV Import when uploading a csv uploads and increases csv count and gives a success message
    1.2 seconds ./spec/system/csv_import_spec.rb:15
  MetadataCloudService it gets called from a rake task is easy to invoke
    1.15 seconds ./spec/services/metadata_cloud_service_spec.rb:36
  MetadataCloudService saving a Voyager record can pull voyager records
    1.06 seconds ./spec/services/metadata_cloud_service_spec.rb:54
  OidImport csv file import can refresh the ParentObjects from the MetadataCloud
    0.98771 seconds ./spec/models/oid_import_spec.rb:33
  ActivityStreamReader daily automated updates can get a page from the MetadataCloud activity stream
    0.83179 seconds ./spec/lib/activity_stream_reader_spec.rb:243
  ActivityStreamReader daily automated updates marks objects as updated in the database
    0.79609 seconds ./spec/lib/activity_stream_reader_spec.rb:247
  ActivityStreamReader daily automated updates can process the partial activity stream if there is a previous successful run
    0.73086 seconds ./spec/lib/activity_stream_reader_spec.rb:266

Top 10 slowest example groups:
  Oid CSV Import
    1.2 seconds average (1.2 seconds / 1 example) ./spec/system/csv_import_spec.rb:4
  METS Xml Imports
    0.63909 seconds average (1.92 seconds / 3 examples) ./spec/system/mets_xml_imports_spec.rb:4
  MetadataCloudService
    0.37767 seconds average (4.15 seconds / 11 examples) ./spec/services/metadata_cloud_service_spec.rb:7
  ParentObject
    0.36382 seconds average (1.82 seconds / 5 examples) ./spec/models/parent_object_solr_spec.rb:5
  ActivityStreamReader
    0.29091 seconds average (4.36 seconds / 15 examples) ./spec/lib/activity_stream_reader_spec.rb:5
  ParentObjects
    0.2814 seconds average (3.66 seconds / 13 examples) ./spec/system/parent_object_spec.rb:4
  OidImport
    0.25979 seconds average (1.04 seconds / 4 examples) ./spec/models/oid_import_spec.rb:5
  IiifManifestFactory
    0.1689 seconds average (3.21 seconds / 19 examples) ./spec/lib/iiif_manifest_factory_spec.rb:5
  ChildObject
    0.13173 seconds average (0.13173 seconds / 1 example) ./spec/models/child_object_spec.rb:5
  /child_objects
    0.12166 seconds average (1.58 seconds / 13 examples) ./spec/requests/child_objects_spec.rb:16

Finished in 27.4 seconds (files took 3.22 seconds to load)
239 examples, 0 failures, 30 pending
```